+++ CHANGES.md
@@ -60,6 +60,7 @@
 * Add anime release groups to add new show options page
 * Add setting "Update shows during hour" to General Config/Misc
 * Add max-width to prevent ui glitch on Pull request and Branch Version selectors on config/General/Advanced and change <input> tags to html5
+* Update Pushbullet notifier (port from midgetspy/sickbeard)
 
 [develop changelog]
 * Change uT params from unicode to str.format as magnet URLs worked but sending files in POST bodies failed